Refactor DI in a cleaner way
