+++ app/src/main/kotlin/com/mpierucci/lasttraintolondon/di/AppComponent.kt
@@ -1,22 +1,17 @@
 package com.mpierucci.lasttraintolondon.di
 
 import com.mpierucci.lasttraintolondon.core.di.CoreComponent
+import com.mpierucci.lasttraintolondon.core.di.scopes.AppScope
 import com.mpierucci.lasttraintolondon.init.AppInitializerModule
 import com.mpierucci.lasttraintolondon.init.CustomApplication
-import com.mpierucci.lasttraintolondon.lines.di.LineStatusesModule
 import com.mpierucci.lasttraintolondon.settings.di.SettingsModule
 import dagger.Component
 
 @Component(
-    modules = [
-        AppInitializerModule::class,
-        LineStatusesModule::class,
-        BindFragmentsModule::class,
-        SettingsModule::class
-    ],
+    modules = [AppInitializerModule::class, SettingsModule::class],
     dependencies = [CoreComponent::class]
 )
-@FeatureScope
+@AppScope
 interface AppComponent {
 
     @Component.Factory
