+++ app/src/main/kotlin/com/mpierucci/lasttraintolondon/di/SingleActivityComponent.kt
@@ -1,9 +1,12 @@
 package com.mpierucci.lasttraintolondon.di
 
+import com.mpierucci.lasttraintolondon.lines.di.LineStatusesModule
 import com.mpierucci.lasttraintolondon.ui.SingleActivity
 import dagger.Subcomponent
 
-@Subcomponent
+@Subcomponent(
+    modules = [LineStatusesModule::class, BindFragmentsModule::class]
+)
 interface SingleActivityComponent {
     fun inject(activity: SingleActivity)
 }
