+++ app/src/main/kotlin/com/mpierucci/lasttraintolondon/ui/SingleActivity.kt
@@ -1,6 +1,5 @@
 package com.mpierucci.lasttraintolondon.ui
 
-import android.content.SharedPreferences
 import android.os.Bundle
 import androidx.appcompat.app.AppCompatActivity
 import androidx.navigation.Navigation
