+++ core/src/main/kotlin/com/mpierucci/lasttraintolondon/core/di/scopes/AppScope.kt
@@ -1,0 +1,7 @@
+package com.mpierucci.lasttraintolondon.core.di.scopes
+
+import javax.inject.Scope
+
+@Scope
+@Retention(AnnotationRetention.RUNTIME)
+annotation class AppScope
